requirements.txt
mitmproxy
obs-websocket-py
